backendP/requirements.txt
django
djangorestframework
drf-spectacular
psycopg2-binary
djangorestframework-simplejwt
django-cors-headers